Fix unit tests to use new saga
